iris classification
The Iris dataset, also known as Fisher's Iris or Anderson's Iris, was
introduced in 1936 by Ronald Fisher in his article "The use of multiple measurements in
taxonomic problems". It serves as an example of the application of linear discriminant analysis.

This dataset consists of 50 samples for each of the three iris species (Iris
setosa, Iris virginica and Iris versicolor). Four characteristics were measured for each
sample: the length and width of the sepals and petals, expressed in centimeters. In
based on the combination of these four variables, Fisher developed an analysis model
linear discriminant allowing species to be distinguished from one another.

In Scikit-learn, SVMs are implemented in the sklearn.svm module. In this code
we will focus on the linear version (Scikit-learn uses the libLinear libraries
and libSVM). We will use the Iris dataset. To be able to display the results, we
will use only the first two attributes (sepal length and width).